Refactor Analysis Models to their own class, also alter names for clarity
