+++ RetireSimple.Engine/Analysis/MonteCarlo.cs
@@ -1,6 +1,6 @@
 using MathNet.Numerics.Distributions;
 
-using RetireSimple.Engine.Data;
+using RetireSimple.Engine.Data.Analysis;
 using RetireSimple.Engine.Data.Investment;
 
 using System.Collections.Concurrent;
