+++ RetireSimple.Engine/Analysis/VehicleDefaultAS.cs
@@ -1,4 +1,4 @@
-using RetireSimple.Engine.Data;
+using RetireSimple.Engine.Data.Analysis;
 using RetireSimple.Engine.Data.Base;
 
 namespace RetireSimple.Engine.Analysis {
