+++ RetireSimple.Engine/Data/Base/Investment.cs
@@ -2,6 +2,7 @@ using Microsoft.EntityFrameworkCore;
 using Microsoft.EntityFrameworkCore.ChangeTracking;
 using Microsoft.EntityFrameworkCore.Metadata.Builders;
 
+using RetireSimple.Engine.Data.Analysis;
 using RetireSimple.Engine.Data.Investment;
 using RetireSimple.Engine.Data.User;
