+++ RetireSimple.Engine/Data/InvestmentVehicle/Vehicle457.cs
@@ -1,4 +1,5 @@
 ﻿using RetireSimple.Engine.Analysis;
+using RetireSimple.Engine.Data.Analysis;
 using RetireSimple.Engine.Data.Base;
 
 namespace RetireSimple.Engine.Data.InvestmentVehicle {
